Grande cambiamento alla gestione dei posti.
